requirements.txt
diffusers==0.30.2
accelerate==0.28.0
numpy==1.23.5
tensorflow==2.12.0
tensorboard==2.12.0
opencv-python==4.9.0.80
soundfile==0.12.1
transformers==4.39.2
huggingface_hub==0.30.2
librosa==0.11.0
einops==0.8.1
gradio==5.24.0

gdown
requests
imageio[ffmpeg]

omegaconf
ffmpeg-python
moviepy

# API server dependencies
fastapi
uvicorn[standard]
python-multipart
pyyaml
